backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
scipy==1.13.0
statsmodels==0.14.2
anthropic==0.28.0
openpyxl==3.1.2
pytest==8.2.0
httpx==0.27.0
google-generativeai==0.7.2
groq==0.9.0